Remove unnecessary files and rename behaviours.py
